Rewrite of the tree manager
This rewrite drops all dependencies on jQuery for the tree manager and
uses modern JavaScript for the drag'n'drop operations. It introduces the
posibility to view the tree as a merged tree of page and media
namespaces (similar to what we do in the ACL manager). This merged view
is often advantageous in wikis where both trees are very similar. A
default view can be set in the configuration but users can switch
between the views on a case by case basis.

This does not change any functionality in any of the other scripts or in
the way how the move is executed. It's a GUI refresh only.

Other parts of the plugin could use a GUI refresh as well (eg. use the
same SVG based icons and drop jQuery dependencies).